LipNet/requirements.txt
absl-py==2.1.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
astor==0.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.11.0
Cython==3.0.11
dlib @ file:///C:/Users/tony_/Downloads/dlib-19.24.4-cp37-cp37m-win_amd64.whl
exceptiongroup==1.2.2
fastapi==0.68.0
fastdtw==0.3.4
Flask==2.2.5
fonttools==4.38.0
gast==0.2.2
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-speech==2.27.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.62.3
grpcio-status==1.62.3
numpy==1.18.5
h11==0.14.0
h5py==2.10.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.0
namex==0.0.8
numpy==1.19.5
opencv-python==3.4.9.33
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
Pillow==9.5.0
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.18
pydantic_core==2.14.6
Pygments==2.18.0
librosa==0.8.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.8
PyYAML==6.0.1
requests==2.31.0
rich==13.8.1
rsa==4.9
scikit-video==1.1.11
scipy==1.2.2
six==1.16.0
sniffio==1.3.1
starlette==0.14.2
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.15.0
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
